+++ src/components/Notification.tsx
@@ -7,11 +7,11 @@ import {stringToColor, timeFormatter} from "@/utils/util"
 import {type FC} from "react"
 import {IoTime, IoCheckbox, IoTrash} from "react-icons/io5"
 import {twMerge} from "tailwind-merge"
-import AvatarImage, {AvatarType} from "../../components/Avatar"
-import Button, {ButtonSize, ButtonVariant} from "../../components/Button"
-import IconButton from "../../components/IconButton"
-import Typography, {TypographyVariant} from "../../components/Typography"
-import {type LocalNotificationData} from "./hooks/useCachedNotifications"
+import AvatarImage, {AvatarType} from "./Avatar"
+import Button, {ButtonSize, ButtonVariant} from "./Button"
+import IconButton from "./IconButton"
+import Typography, {TypographyVariant} from "./Typography"
+import {type LocalNotificationData} from "../containers/NavigationSection/hooks/useCachedNotifications"
 
 export interface NotificationProps extends LocalNotificationData {
   onRequestChanges: () => void
